[CDAP-17446] Replication Detail Page Redesign
