+++ cdap-ui/app/cdap/components/Replicator/Detail/Monitoring/SelectedTable/index.tsx
