+++ cdap-ui/app/cdap/components/Replicator/Detail/Monitoring/TimePeriodSelection/index.tsx
@@ -1,0 +1,39 @@
+/*
+ * Copyright © 2020 Cask Data, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not
+ * use this file except in compliance with the License. You may obtain a copy of
+ * the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations under
+ * the License.
+ */
+
+import React from 'react';
+import withStyles, { WithStyles, StyleRules } from '@material-ui/core/styles/withStyles';
+import TimePeriodDropdown from 'components/Replicator/Detail/TimePeriodDropdown';
+
+const styles = (): StyleRules => {
+  return {
+    applyFont: {
+      fontSize: '22px',
+    },
+  };
+};
+
+const TimePeriodSelectionView: React.FC<WithStyles<typeof styles>> = ({ classes }) => {
+  return (
+    <div>
+      <span className={classes.applyFont}>Replication pipeline activities - View</span>
+      <TimePeriodDropdown classes={{ selectRoot: classes.applyFont }} />
+    </div>
+  );
+};
+
+const TimePeriodSelection = withStyles(styles)(TimePeriodSelectionView);
+export default TimePeriodSelection;
