+++ cdap-ui/app/cdap/components/Table/TableColumnGroup.tsx
@@ -1,0 +1,49 @@
+/*
+ * Copyright © 2020 Cask Data, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not
+ * use this file except in compliance with the License. You may obtain a copy of
+ * the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations under
+ * the License.
+ */
+
+import React from 'react';
+import withStyles, { WithStyles, StyleRules } from '@material-ui/core/styles/withStyles';
+
+const styles = (theme): StyleRules => {
+  return {
+    columnGroups: {
+      backgroundColor: theme.palette.white[50],
+      display: 'grid',
+    },
+  };
+};
+
+interface ITableColumnGroupProps extends WithStyles<typeof styles> {
+  columnTemplate?: string;
+}
+
+const TableColumnGroupView: React.FC<React.PropsWithChildren<ITableColumnGroupProps>> = ({
+  classes,
+  children,
+  columnTemplate,
+}) => {
+  const style = {
+    gridTemplateColumns: columnTemplate,
+  };
+  return (
+    <div className={classes.columnGroups} style={style}>
+      {children}
+    </div>
+  );
+};
+
+const TableColumnGroup = withStyles(styles)(TableColumnGroupView);
+export default TableColumnGroup;
